requirement.txt
Django==1.11.15
django-bootstrap3==10.0.1
django-rest-framework==0.1.0
djangorestframework==3.8.2
pkg-resources==0.0.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pytz==2018.5
